Moved `RelatedObjectUrlSlug` enum into `constants` module
This is done so that this enum can be imported/used in integration
tests without having to import `server/data/incidentReportingApi`
which imports `logger`/`bunyan` which can't be used anymore in
Cypress tests because of the removed Node's `util` polyfill.

This makes sense anyway as the enum is used as a collection
of constants anyway.
